==> notebooks/01_academics-on-mastodon.ipynb <==
%pip install parse
%pip install GitPython

%pip install requests
%pip install pandas

%pip install Office365-REST-Python-Client

%pip install beautifulsoup4

import os
import glob
import parse
import shutil
import requests
import pandas as pd
from git import Repo
from git import GitError
from io import StringIO

source = "https://github.com/nathanlesage/academics-on-mastodon/blob/main/README.md"


!wget https://raw.githubusercontent.com/nathanlesage/academics-on-mastodon/main/README.md -O 'academics-source.md'

file = open("temp/academics-source.md").readlines()


format_string = "- [{}]({}){}"
sources = []

islist = False
for line in file:
    if line == "## Curated Account Lists\n":
        islist = True
    elif islist and line[:2] == "##":
        islist = False
    if islist and line[0] == "-":
        parsed = parse.parse(format_string, line)
        field_csv = (parsed[0], parsed[1])
        sources.append(field_csv)


def get_csv(txt):
    output = []
    for line in txt.split('\n'):
        if line.startswith('@') and len(line) > 2:
            parsed = line[:-2].split(' (') #redo
            if len(parsed) > 1:
                output.append((parsed[0], parsed[1]))
    return output


def get_drive_file(source):
    field = source[0]
    url = source[1]
    str_format = "https://docs.google.com/{}/d/{}/{:z}"
    parsed = parse.parse(str_format, url, {"z": zero_or_more_string})
    doc_type = parsed[0]
    code = parsed[1]

    if doc_type == "spreadsheets":
        result = requests.get(
            "https://docs.google.com/{}/d/{}/export?format=csv".format(
                doc_type, code
            )
        )
        csv_string = result.text
        
    elif doc_type == "document":
        export_url = "https://docs.google.com/{}/d/{}/export?format=txt".format(
            doc_type, code
        )
        result = requests.get(export_url)
        csv_string = result.text
        
    return csv_string


def get_github_file(source):

    field = source[0]

    str_format = 'https://{}.github.io/{}/'
    parsed = parse.parse(str_format, source[1])
    user_git = parsed[0]
    repo_name = parsed[1]

    path_file = 'temp/' + repo_name + '_copy'
    if not os.path.isdir(path_file):
        try:
            Repo.clone_from('https://github.com/{}/{}'.format(user_git, repo_name), path_file)
        except GitError as error:
            print("Error:\n user: {}, repo: {}, msg: {}".format(user_git, repo_name, error))
            return None
    

    files = glob.glob(path_file + '/resources/*.csv')
    csv_str = ''
    for file in files:
        csv_str += open(file, encoding="utf8").read() + '\n\n'

    '''
    try:
        shutil.rmtree(path_file)
    except:
        print('permission denied to del folder temp/{}'.format(repo_name))
    
    '''
    return csv_str


def zero_or_more_string(text):
    return text


zero_or_more_string.pattern = r".*"


data = []
for source in sources:
    field = source[0].replace("/", " ")
    if "docs.google.com" in source[1]:
        txt = get_drive_file(source)
        data.append((field, txt))
    elif "sharepoint" in source[1]:
        #         https://faun.pub/quick-etl-with-python-part-1-download-files-from-sharepoint-online-40bf23711662
        pass
        
    elif "github" in source[1]:
        txt = get_github_file(source)
        if df != None:
            data.append((field, txt))
    


for field, field_csv in data:
    if field_csv:
        output = open('csv/scholars_{}.csv'.format(field), "w", encoding="utf-8")
        output.write(field_csv)
        output.close()

https://github.com/fivestone/fedi-anthropologists/archive/refs/heads/main.zip
    
    git@github.com:fivestone/fedi-anthropologists.git
