==> notebooks/03_mastodon-instances-list.ipynb <==
import json
import requests

url = 'https://instances.social/api/1.0/instances/list'

token = open('mastodon-instances-token').read().split('\n')[-1]
print(token)

r = requests.get(url, headers={'Authorization': "Bearer {}".format(token), 'count': '0', 'include_down': 'false'})

r.status_code

instances = json.loads(r.text)

for instance in instances['instances']:
    print(instance['id'], instance['name'], instance['added_at'])

instance_list = []
c = 1
while True:
    c += 1
    print(c, end='\r')
    test = json.loads(r.text)

    for instance in test['instances']:
        instance_list.append((instance['id'], instance['name']))

    next = test['pagination']
    next_id = next['next_id']

    if next_id == None:
        break
    if c > 900:
        print('break with c')
        break
    r = requests.get(url, headers={'Authorization': "Bearer {}".format(token)}, params={'min_id': next_id})

# for instance in test['instances']:
#     instance_list.append((instance['id'], instance['name']))

len(instance_list)

out = open('instance_list.tsv', 'w', encoding='utf-8')
for instance in instance_list:
    out.write("{}\t{}\n".format(*instance))
out.close()
