==> notebooks/04_app_for_each_instance.ipynb <==
import glob
import time
import pandas as pd
from mastodon import Mastodon


files = glob.glob('../data/users_followers_*_2023-05-02.csv')
all_users = []
for file in files:
    users = pd.read_csv(file, sep='\t', names=['user', 'instance', 'followers', 'field'], encoding_errors='replace', on_bad_lines='warn')
    all_users.append(users)

all_users = pd.concat(all_users)

instances = users['instance'].unique()

instances_valids = open('instance_list.tsv', encoding='utf-8').read()
def isinstance(name):
    return name in instances_valids

for instance_name in instances:

    if not isinstance(instance_name):
        print(instance_name)
        continue

    Mastodon.create_app(
        'pytooterapp',
        api_base_url = 'https://{}'.format(instance_name),
        to_file = 'temp/mastodon_app_key_{}.secret'.format(instance_name)
    )
    time.sleep(3)
