==> notebooks/05_mastodon-users-toots.ipynb <==
import time
import glob
import json
import os.path
import requests
import pandas as pd
from  datetime import date

today = date.today()

files = glob.glob('../data/users_followers_*_2023-05-02.csv')
all_users = []
for file in files:
    users = pd.read_csv(file, sep='\t', names=['user', 'instance', 'followers', 'field'], encoding_errors='replace', on_bad_lines='warn')
    all_users.append(users)

all_users = pd.concat(all_users)
print(all_users.head())

import requests
from requests.adapters import HTTPAdapter
from urllib3.util.retry import Retry

session = requests.Session()
retry = Retry(connect=3, backoff_factor=0.5)
adapter = HTTPAdapter(max_retries=retry)
session.mount('http://', adapter)
session.mount('https://', adapter)


def get_userid(user_name, user_server, token_access):
    try:
        url = 'https://{}/api/v1/accounts/lookup/'.format(user_server)
        headers = {
            'Authorization' : 'Bearer {}'.format(token_access)
        }
        params = {
            'acct' : '{}@{}'.format(user_name, user_server)
        }
        r = session.get(url, headers=headers, params=params)

        user = json.loads(r.text)
        if 'id' in user:
            return user["id"]
        elif 'user' in user:
            return user["user"]
    except Exception as error:
        print(error)
        return None

mastodon_app = "acmbrito_app_python_secret"
client_id, client_secret, access_token, api_base_url = open(mastodon_app).read().split()

get_userid('keirol', 'mastodon.social', access_token)

date_limit = date(2022, 9, 1)

import util
def get_instance_token(instance):
    return util.get_instance_token(instance)

# toots

# source to collect user toots: https://jrashford.com/2023/02/13/how-to-scrape-mastodon-timelines-using-python-and-pandas/
# API code: https://mastodonpy.readthedocs.io/en/1.8.1/_modules/mastodon/timeline.html?highlight=users%20in%20a%20given%20list
# https://docs.joinmastodon.org/methods/search/#v1

def user_toots(instance, user_id, access_token):
    URL = 'https://{}/api/v1/accounts/{}/statuses'.format(instance, user_id)
    user_toots = []
    if access_token:
        headers = {
            'Authorization' : 'Bearer {}'.format(access_token)
        }
        
        user_toots = []
        params = {}
        while True:
            #time.sleep(1)
            r = requests.get(URL, headers=headers, params=params, timeout=60)
            toots = json.loads(r.text)
            user_toots += toots
            if len(toots) == 0:
                break
            if 'error' in toots:
                print(user_id, toots)
                break
            max_id = toots[-1]['id']
            params = {'max_id' : max_id}
            
            date_str = toots[-1]['created_at'].split('T')[0]
            last_date = date.fromisoformat(date_str)
            if last_date < date_limit:
                break

    return user_toots

valid_fields = ['id',
'created_at',
'in_reply_to_account_id',
'uri',
'replies_count',
'reblogs_count',
'favourites_count',
'content'
]

files = glob.glob('temp/users_toots_2023-0*.tsv')
users_already_collected = []
invalid_users = set()
for file in files:
    invalid_users |= set(pd.read_csv(file, sep='\t')['user_id'])

print(len(invalid_users))

outputs = []
size = 0
user_id_map = pd.read_csv('user_id_list.tsv', sep='\t')
user_id_map = user_id_map.drop_duplicates(subset=['user', 'instance'])
user_id_map = user_id_map.dropna()

for _, user in all_users.iterrows():
    
    user_id = user_id_map[(user_id_map['user'] == user['user']) & (user_id_map['instance'] == user['instance'])]
    if len(user_id) >= 1:
        user_id = user_id['user_id'].values[0]
        
    else:
        continue

    if not user_id:
        print(user)
        continue

    if not user_id.isnumeric():
        continue

    if int(user_id) in invalid_users:
        continue
    
    instance_token = get_instance_token(user['instance'])
    if not instance_token:
        continue

    toots = user_toots(user['instance'], user_id, instance_token)
    if len(toots) > 0:
        toots_pd = pd.json_normalize(toots)
        toots_pd = toots_pd.loc[:, valid_fields]
        toots_pd['user_id'] = user_id
        size += len(toots_pd)
        outputs.append(toots_pd)

    if size > 10000:
        data_toots = pd.concat(outputs)
        data_toots.to_csv('temp/users_toots_{}_{}.tsv'.format(today, user_id), sep='\t')
        outputs = []
        size = 0

data_toots = pd.concat(outputs)
data_toots.to_csv('temp/users_toots_{}_{}.tsv'.format(today, user_id), sep='\t')

#TODO to select instances related to scholars/science
#questions: search for scholars on mastodon.social? (the "main" instance)

# o tamanho de um subconjunto para selecionar?????
# fediscience (onde todos os usuarios sao academicos) como primeiro subset
# como expandir isso para o mais instancias?
# 

ids = []
i = 0
'''
for _, user in all_users.iterrows():
    instance_token = get_instance_token(user['instance'])
    user_id = get_userid(user['user'], user['instance'], instance_token)
    ids.append((user['user'], user['instance'], user_id))
    i += 1
    if i % 300 == 0:
        print(i, end='\r')
        pd.DataFrame(ids, columns=['user', 'instance', 'user_id']).to_csv('user_id_list.tsv', sep='\t')
'''


filename = '../data/mastodon_users_wOpenAlex.csv'
openalex_masto = pd.read_csv(filename, sep=',')
print(openalex_masto.columns)
openalex_masto = openalex_masto.drop(columns='Unnamed: 0').drop_duplicates()
print(openalex_masto.size)
openalex_masto.head()

ids = []
i = 0

for _, user in openalex_masto.iterrows():
    mastodon_name = user['mastodon_name'].split("@")[1:]
    user_name = mastodon_name[0]
    instance_name = mastodon_name[1]
    instance_token = get_instance_token(instance_name)
    user_id = get_userid(user_name, instance_name, instance_token)
    ids.append((user_name, instance_name, user_id))
    i += 1
    if i % 100 == 0:
        print(i, end='\r')
        pd.DataFrame(ids, columns=['user', 'instance', 'user_id']).to_csv('user111_id_list_kunilist_{}.tsv'.format(today), sep='\t')


## kuni list of users -> toots collected in 09-07-2023 (dd/mm/yyyy)

outputs = []
size = 0
user_id_map = pd.read_csv('user111_id_list_kunilist_2023-06-30.tsv', sep='\t', dtype=object)
user_id_map = user_id_map.drop_duplicates(subset=['user', 'instance'])
user_id_map = user_id_map.dropna()
#user_id_map = user_id_map.astype({'user_id':'int'})
print(user_id_map.head())

for _, user in user_id_map.iterrows():
    '''print(user)
    item = user['mastodon_name'].split('@')
    
    user = {
        'user': item[1],
        'instance': item[2]
    }'''

    user_id = user_id_map[(user_id_map['user'] == user['user']) & (user_id_map['instance'] == user['instance'])]
    if len(user_id) >= 1:
        user_id = user_id['user_id'].values[0]
    else:
        continue

    if not user_id:
        print(user)
        continue

    if not user_id.isnumeric():
        continue

    # if int(user_id) in invalid_users:
    #    continue
    
    instance_token = get_instance_token(user['instance'])
    if not instance_token:
        continue

    toots = user_toots(user['instance'], user_id, instance_token)
    if len(toots) > 0:
        toots_pd = pd.json_normalize(toots)
        toots_pd = toots_pd.loc[:, valid_fields]
        toots_pd['mastodon_name'] = "{}@{}".format(user['user'], user['instance']) 
        size += len(toots_pd)
        outputs.append(toots_pd)

    if size > 10000:
        data_toots = pd.concat(outputs)
        data_toots.to_csv('temp/users_toots_{}_{}.tsv'.format(today, user_id), sep='\t')
        outputs = []
        size = 0

data_toots = pd.concat(outputs)
data_toots.to_csv('temp/users_toots_{}_{}.tsv'.format(today, user_id), sep='\t')

## kuni list of users -> toots collected in 09-07-2023 (dd/mm/yyyy)
