==> notebooks/06_mastodon_users_to_twitter_handles.ipynb <==
import re
import glob
import pandas as pd
import random
import datetime

today = datetime.date.today()


handle = "twitter"

files = glob.glob("csv/*.csv") + glob.glob("csv/others/*.csv")
mastodon_users = dict()
twitter = []
for file in files:
    df = pd.read_csv(file, on_bad_lines="warn")
    mastodon_col = ""
    for col in df.columns:
        if "mastodon" in col.lower() or "account" in col.lower():
            mastodon_col = col
            break
    for col in df.columns:
        if handle in col.lower():
            scholars = df[[mastodon_col, col]]
            scholars = scholars.rename(
                columns={mastodon_col: "mastodon", col: "twitter"}
            )
            twitter.append(scholars)

df = pd.concat(twitter)
df = df[df["twitter"] != "no"]
df = df[df["twitter"] != "No"]
df = df.dropna()
df.to_csv("../data/mastodon2twitter_{}.csv".format(today), sep="\t", index=False)

file = sorted(glob.glob('../data/mastodon2twitter*.csv'))[-1] # select the most recent version
df = pd.read_csv(file, sep='\t', encoding="utf8")
df.head()

def is_valid_twitter_handler(sequence):
    if ' ' in sequence:
        return False
    match = re.fullmatch('^@(\w{1,15})$', sequence)
    return match

twitter_handles = []
for row in df['twitter']:
    if is_valid_twitter_handler(row):
        twitter_handles.append(row)

# sample5k = random.sample(twitter_handles, 5000)

output = open('../data/valid_twitter_handles_{}_complete.txt'.format(today), 'w', encoding='utf-8')
for line in twitter_handles:
    output.write("{}\n".format(line))
output.close()

!wget https://tavern.osome.iu.edu/jobs/2023-05-04/d29d6f7d-7d53-4968-a6b4-09c1c4be4570/ -O
