==> notebooks/07_mastodon-overview-data.ipynb <==
# Academics on Mastodon

https://github.com/nathanlesage/academics-on-mastodon

import glob
import pandas as pd


files = glob.glob('temp/users_toots_2023-06-30*.tsv')

toots = []
for file in files:
    current_toots = pd.read_csv(file, sep='\t')
    toots.append(current_toots)

toots = pd.concat(toots)
print(toots.head())

## Total number of toots collected: 1,042,585
### Kuni list of scholars: 186,941

Since: 2022-09-01

Metadata:
- id
- created_at
- in_reply_to_account_id
- uri
- replies_count
- reblogs_count
- favourites_count
- content
- user_id

print(len(toots))

## Total number of users (active publishing on Mastodon): 4478
### Kuni list of scholars: 1392 of 4599
## Inactive active users: 6748

print(len(toots['user_id'].unique()))

toots.columns

files = glob.glob('../data/users_followers_*_2023-05-02.csv')
all_users = []
for file in files:
    users = pd.read_csv(file, sep='\t', names=['user', 'instance', 'followers', 'field'], encoding_errors='replace', on_bad_lines='warn')
    all_users.append(users)

all_users = pd.concat(all_users)
print(all_users.head())
print(len(all_users))

user_id_map = pd.read_csv('user_id_list.tsv', sep='\t')
user_id_map = user_id_map.drop_duplicates(subset=['user', 'instance'])

users = pd.merge(user_id_map, all_users, on=['user', 'instance'])
print(users.head())

users = users.astype({'user_id': 'object'})
toots = toots.astype({'user_id': str})

toots.head()

users_toots = pd.merge(users, toots, on=['user_id'])
print(users_toots.head())

## Number of toots published by "topic" (author's field of research):

users_toots.groupby('field')['id'].count().sort_values(ascending=False)

# missing users fields

invalid = set(toots['user_id'].unique())
missing_users = users[~users['user_id'].isin(invalid)]
print(len(missing_users))

missing_count = missing_users.groupby('field')['user_id'].count().sort_values(ascending=False)

users_by_field = users.groupby('field')['user_id'].count().sort_values(ascending=False)
users_by_field.head()

## Percentage of users missing (or not active on Mastodon) by field of study

users_overview = pd.merge(users_by_field, missing_count, on='field')
users_overview['percentage'] = users_overview.apply(lambda row: row['user_id_y']/row['user_id_x']*100, axis=1)
users_overview.sort_values('percentage', ascending=True).head(20)

users_toots_description = users_toots.groupby(['user', 'instance'])['id'].count().sort_values(ascending=False)
display(users_toots_description.describe())

## questions
- time window (collect toots since when)
- number of inactive users problem 
- authors with 2 or more fields of study


### about mastodon data
- collect toots from instance to collect their authors (what about users that never published)
- criterias to select mastodon users: academic instances, any popular instance (for example mastodon.social) or usernames mentioned on tweets
- frequency/engagement

TODO LIST
- blue sky
- diagram of users from source x to source y
- use de mastodon users collected by kuni
- tweets about mastodon vs tweets general over time

https://www.nature.com/articles/504211a

https://www.nature.com/articles/550029a

https://www.nature.com/articles/d41586-023-00486-3

https://yyiki.org/wiki/GitHub/Workflow/
