==> notebooks/08_academics_table_scholars_by_fos.ipynb <==
# TODO table about the users collected and their labels (instances)

import glob
import pandas as pd

!pip install Jinja2

user_followers = glob.glob('../data/users_followers_*2023-05-02.csv')
full_data = []
for file in user_followers:
    print(file)
    data = pd.read_csv(file, sep='\t', header=None, encoding="utf8", encoding_errors='ignore', on_bad_lines='skip')
    data.columns = ['user', 'instance', 'followers', 'field']
    full_data.append(data)


full_data = pd.concat(full_data)

str_table = full_data.groupby('field')['field'].count().to_latex()
output = open('str_table_fields', 'w')
output.write(str_table)
output.close()

len(full_data)
