==> notebooks/09_preprocess_openalex_kuni.ipynb <==
import ast
import glob
import numpy as np
import pandas as pd

files = glob.glob('temp/mastodon_users_papers_*.tsv')

files[:3]

len(files)

data = []
for file in files:
    data.append(pd.read_csv(file, sep='\t'))
    
data = pd.concat(data)

len(data)

data = data.drop('Unnamed: 0', axis=1)
data.head()

from collections import defaultdict

authors_year = defaultdict(lambda:[])
for idx, row in data.iterrows():
    authors = ast.literal_eval(row['authorships:author:id'])
    year = row['publication_year']
    for author in authors:
        authors_year[author].append(year)

len(authors_year)

data = pd.read_csv('kuni_twitter_open_alex_mstd_07-18-2023.tsv', sep='\t')
data = data.drop(columns=['Unnamed: 0'], axis=1)

# ntoots = pd.read_csv('kuni_ntoots.tsv', sep='\t')
# data = data.merge(ntoots, on='mastodon_name')
data.head()

BIRTH = []
NPAPERS = []
for idx, row in data.iterrows():
    author_openalex_id = int(row['OpenAlex_account'][22:])
    papers = authors_year[author_openalex_id]
    if len(papers) > 0:
        birth = min(papers)
        npapers = len(papers)
    else:
        birth = None
        npapers = None
    
    BIRTH.append(birth)
    NPAPERS.append(npapers)

data['birth_openalex'] = BIRTH
data['npapers_openalex'] = NPAPERS

for k,v in authors_year.items():
    print(k, v)

data.to_csv('kuni_authors_with_open_alex.tsv', sep='\t')
