==> notebooks/14_users_folllowing.ipynb <==
import glob
import pandas as pd

files = glob.glob('temp/users_following_kuni_*.tsv')
data = []
for file in files:
    temp = pd.read_csv(file, sep='\t')
    data.append(temp)

data = pd.concat(data)
data.columns = ['idx', 'source_acct', 'id', 'target_acct', 'locked', 'bot', 'discoverable', 'created_at', 'note', 'followers_count', 'following_count', 'statuses_count', 'last_status_at', 'fields']
print(data.head())

data.to_csv('users_mstd_kuni_following.tsv', sep='\t', index=False)

profiles = pd.read_csv('kuni_twitter_open_alex_mstd_07-18-2023.tsv', sep='\t')
profiles = profiles.drop(columns=['Unnamed: 0'], axis=1)
print(len(profiles))
profiles.head()

profile_following = profiles[['mastodon_name', 'following']]

following_from_api = data.groupby('source_acct').count()['idx']

check = profile_following.merge(following_from_api, left_on='mastodon_name', right_on='source_acct')

check
