==> notebooks/15_concepts_unification.ipynb <==
!pip install jellyfish

import glob
import jellyfish
import numpy as np
import pandas as pd

concepts = pd.read_csv('OpenAlex concepts in use (17 August 2022) - concepts.tsv', sep='\t')
concepts.head()

concepts_list = concepts['normalized_name'].values
print(len(concepts_list))
concepts_list = concepts_list[~pd.isnull(concepts_list)]
print(len(concepts_list))

files = glob.glob("csv/*.csv") + glob.glob('csv/others/*.csv')
print(files[:4])

fields_github = dict()
for file in files:
    name = file.split('/')[-1]
    if 'scholars_' in name:
        name = name[9:]
    name = name[:-4]
    fields_github[file] = name

fields_github

for k,v in fields_github.items():
    print(k)
    dists = []
    for current in concepts_list:
        dists.append(jellyfish.levenshtein_distance(v, current))
    idxs = np.argsort(dists)
    for idx in idxs[:3]:
        print(concepts_list[idx])
    print()

content = open('concepts_unification.txt').read().split('\n\n')
mapconcepts = dict()
for row in content:
    temp = row.split('\n')
    temp2 = []
    for item in temp:
        temp2.append(item.split(" (")[0])

    mapconcepts[temp[0]] = tuple(temp2[1:])
    

mapconcepts

import ast
import math
from collections import defaultdict
    
def level_0_concept(concepts, current):
    if type(current['parent_ids']) == str:
        a = []
        fields = current['parent_ids'].split(', ')
        for current_id in fields:
            a += level_0_concept(concepts, concepts[concepts['openalex_id'] == current_id.lower()].iloc[0])
        return a
    else:
        return [current['normalized_name']]

mapconcepts_0level = defaultdict(lambda: []) # 19 giant -> subtopics
for file, fields in mapconcepts.items():
    for field in fields:
        print(field)
        temp = level_0_concept(concepts, concepts[concepts['normalized_name'] == field].iloc[0])
        for level0 in temp:
            mapconcepts_0level[level0].append(field)


map_csv = []
for k, v in mapconcepts_0level.items():
    v = set(v)
    for subtopic in v:
        map_csv.append((k, subtopic))

pd.DataFrame(map_csv).to_csv('topic2subtopic.csv')
