==> notebooks/16_reply_net.ipynb <==
import openai
import numpy as np
import igraph as ig
import pandas as pd
import xnetwork as xnet
from datetime import date
from bs4 import BeautifulSoup

profiles = pd.read_csv('mentioned_profiles_kuni_2023-08-07 01:20:19.437081.tsv', sep='\t')
print(profiles.columns)

notes = profiles[['username', 'note']]
notes = notes.fillna('')
concepts = pd.read_csv('OpenAlex concepts in use (17 August 2022) - concepts.tsv', sep='\t')
concepts_lvl = concepts[concepts['level'] < 2]

count = 0
concepts_lvl = concepts_lvl['normalized_name'].dropna().values
mentions_concepts = []
for idx, row in notes.iterrows():
    description = BeautifulSoup(row['note']).text.lower()
    scholar_interests = []
    for concept in concepts_lvl:
        if concept in description:
            scholar_interests.append(concept)
    if len(scholar_interests) > 0:
        count += 1
    mentions_concepts.append(scholar_interests)
print(count, len(mentions_concepts))

concepts_map = dict()

concepts_lvl = concepts[concepts['level'] < 2]
for idx, row in concepts_lvl.iterrows():
    if row['level'] == 1:
        concepts_map[row['normalized_name']] = row['parent_display_names'].lower().split(', ')
    else:
        concepts_map[row['normalized_name']] = [row['normalized_name']]

new_name_concepts = []
for items in mentions_concepts:
    new_name = []
    for item in items:
        new_name += concepts_map[item]
    if len(new_name) > 0:
        new_name_concepts.append(max(new_name))
    else:
        new_name_concepts.append('unknown')

profiles['x_concepts'] = new_name_concepts

profiles = profiles[['username', 'display_name', 'followers_count', 'following_count', 'url', 'x_concepts']]

profiles.head()

profiles = profiles.dropna(subset='username')

def get_instance(row):
    return "{}".format(row['url'][len("https://"):].split('/')[0])
    
def get_mastodon_name(row):
    instance = get_instance(row)
    return "{}@{}".format(row['username'], instance)

profiles['mastodon_name'] = profiles.apply(get_mastodon_name, axis=1)
profiles['instance'] = profiles.apply(get_instance, axis=1)

profiles.head()

mentions = pd.read_csv('kuni_users_mentions.tsv', sep='\t')

pairs = mentions[['toot_author', 'mentioned_user_instance']]
pairs.values

unique = np.unique(pairs.values)

fields_of_study = pd.read_csv('mastodon_users_wOpenAlex2.csv')[['mastodon_name', 'OpenAlex_account', 'x_concepts']]
fields_of_study['x_concepts'] = fields_of_study['x_concepts'].fillna('unknown')


def edit_mastodon_name(row):
    temp = row['mastodon_name']
    if temp.startswith("@"):
        temp = temp[1:]
    return temp


fields_of_study['mastodon_name'] = fields_of_study.apply(edit_mastodon_name, axis=1)

fields_of_study.head()

from ast import literal_eval

def get_concept(row):
    if row['x_concepts'] == 'unknown':
        return 'unknown'
    list_of_concepts = literal_eval(row['x_concepts'])
    concepts = []
    if len(list_of_concepts) == 0:
        return 'unknown'
    else:
        for concept in list_of_concepts:
            if concept['level'] == 0:
                return concept['display_name']

fields_of_study['x_concepts'] = fields_of_study.apply(get_concept, axis=1)

fields_of_study['instance'] = fields_of_study.apply(lambda row: row['mastodon_name'].split("@")[1], axis=1)
fields_of_study.head()

user_profiles = pd.concat([fields_of_study, profiles])

user_profiles.head(30)

print(len(user_profiles))
user_profiles = user_profiles.drop_duplicates(subset='mastodon_name')
print(len(user_profiles))

g_mentions = ig.Graph(directed=True)
g_mentions.add_vertices(len(user_profiles))
g_mentions.vs['name'] = user_profiles['mastodon_name'].values

valid_pairs = []
for idx, pair in pairs.iterrows():
    if pair.toot_author in g_mentions.vs['name'] and pair.mentioned_user_instance in g_mentions.vs['name']:
        valid_pairs.append(pair)

g_mentions.add_edges(valid_pairs)

g_mentions.vs['instances'] = user_profiles['instance'].values

g_mentions.vs['concepts'] = concepts

count

xnet.igraph2xnet(g_mentions, 'g_mentions.xnet')
