==> notebooks/16_reply_net_github.ipynb <==
import glob
import pandas as pd

user_id_map = pd.read_csv('user_github_id_list.tsv', sep='\t', dtype='object').dropna()
user_id_map = user_id_map.drop_duplicates(subset=['user_instance', 'user_id'])

user_id_map.head()

import numpy as np

u,c = np.unique(user_id_map['user_id'].values, return_counts=True)

len(user_id_map)

user_id_map[user_id_map['user_id'] == '109321922930929886']

files = glob.glob("temp/users_toots_github_*.tsv")

len(files)

all_data = []
for file in files:
    data = pd.read_csv(file, sep='\t', dtype='object')
    all_data.append(data)
    
data = pd.concat(all_data)

data.columns

data_with_id = data.merge(user_id_map, on='user_id')

data_with_id = data_with_id.drop(columns=['Unnamed: 0_x', 'Unnamed: 0_y'])
data_with_id.head()

github_users = pd.read_csv('mastodon_profiles_github.tsv', sep='\t')
github_users.columns = ['index', 'user_instance', 'display_name', 'twitter', 'followers', 'following', 'created_at', 'bio', 'field']
github_users = github_users.drop(columns=['field', 'index'])

fields_users = pd.read_csv('mastodon_profiles_fields.tsv', sep='\t')
fields_users.columns = ['index', 'user_instance', 'field_file']
fields_users = fields_users.drop(columns=['index'])

github_users = github_users.merge(fields_users, on='user_instance', how='left')
github_users = github_users.dropna(subset='field_file')

print(github_users.head())

subtopics_map = pd.read_csv('topic2subtopic.csv')
subtopics_map.columns = ['idx', 'level0', 'subtopic']
subtopics_map.head()

content = open('concepts_unification.txt').read().split('\n\n')
mapconcepts = dict()
for row in content:
    temp = row.split('\n')
    temp2 = []
    for item in temp:
        temp2.append(item.split(" (")[0])

    mapconcepts[temp[0]] = tuple(temp2[1:])

fields_unification = []
for idx, row in github_users.iterrows():
    temp = mapconcepts[row['field_file']]
    fields_unification.append(temp)

fields_unification[:5]

users_fields = []
for field in fields_unification:
    field = field[0]
    temp = subtopics_map[subtopics_map['subtopic'] == field].iloc[0]['level0']
    users_fields.append(temp)
    #como lidar com mais de uma área?

github_users['field_level0'] = users_fields

mentions = pd.read_csv('github_users_mentions.tsv', sep='\t')

mentions.columns

nodes = set()
edges = []
for idx, row in mentions.iterrows():
    source = row['toot_author']
    target = row['mentioned_user_instance']
    edges.append((source, target))
    nodes.add(source)
    nodes.add(target)

len(nodes),len(edges)

nodes = list(nodes)
instances = [node.split('@')[1] for node in nodes]

import igraph as ig
g_mentions = ig.Graph(directed=True)

g_mentions.add_vertices(len(nodes))
g_mentions.vs['name'] = nodes
g_mentions.vs['instance'] = instances
g_mentions.add_edges(edges)
g_mentions.simplify()

fields = []
for node in nodes:
    temp = github_users[github_users['user_instance'] == node]
    if len(temp) > 0:
        f = temp.iloc[0]['field_level0']
    else:
        f = 'unknown'
    fields.append(f)

g_mentions.vs['field'] = fields

g_mentions.vs['in_degree'] = g_mentions.degree(mode='IN')

import xnetwork as xnet

xnet.igraph2xnet(g_mentions, 'g_mentions_github_simplify.xnet')

len(set(edges))

g_mentions.ecount()
